+++ scripts/changelog.ts
@@ -196,8 +196,7 @@ async function exec(): Promise<void> {
         '**all**. Added plugin icons and descriptions',
         '**plugin-text**. Added Markdownshortcuts for Lists, Headings and Blockquotes.',
         '**core**. Pass plugin props and name of parents to childs too',
-        '**plugin-text**. Leave nested Text-Plugins on second Enter key press',
-        '**plugin-text**. Poweruser-Feature: Add new plugins with /[plugintitle] in editmode'
+        '**plugin-text**. Leave nested Text-Plugins on second Enter key press'
       ],
       changed: [
         '**plugin-text**. Hide placeholder in render mode',
@@ -217,7 +216,8 @@ async function exec(): Promise<void> {
     {
       added: [
         '**plugin-files**. Add plugin for file uploads',
-        '**plugin-rows**. Pass `renderIntoExtendedSettings` and `PrimarySettings` as props to children'
+        '**plugin-rows**. Pass `renderIntoExtendedSettings` and `PrimarySettings` as props to children',
+        '**plugin-text**. Poweruser-Feature: Add new plugins with /[plugintitle] in editmode'
       ],
       changed: [
         '**plugin-rows**. Adapt styles and controls implemented by schul-cloud, including drag&drop',