+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - **plugin-files**. Add plugin for file uploads
 - **plugin-rows**. Pass `renderIntoExtendedSettings` and `PrimarySettings` as props to children
+- **plugin-text**. Poweruser-Feature: Add new plugins with /[plugintitle] in editmode
 
 ### Changed
 
@@ -46,7 +47,6 @@ All notable changes to this project will be documented in this file.
 - **plugin-text**. Added Markdownshortcuts for Lists, Headings and Blockquotes.
 - **core**. Pass plugin props and name of parents to childs too
 - **plugin-text**. Leave nested Text-Plugins on second Enter key press
-- **plugin-text**. Poweruser-Feature: Add new plugins with /[plugintitle] in editmode
 
 ### Changed